git-svn-id: svn://192.168.1.10/public/CoffeeMud@11118 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Behaviors/Mobile.java
@@ -81,8 +81,8 @@ public class Mobile extends ActiveTicker implements MobileBehavior
 		{
 			if(currentRoom==null) return false;
 			if(leashHash==null)    leashHash=new Hashtable();
-			Integer DISTNOW=(Integer)leashHash.get(currentRoom);
-			Integer DISTLATER=(Integer)leashHash.get(newRoom);
+			Integer DISTNOW=leashHash.get(currentRoom);
+			Integer DISTLATER=leashHash.get(newRoom);
 			if(DISTNOW==null)
 			{
 				DISTNOW=Integer.valueOf(0);
